backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.6
sqlmodel==0.0.22
python-multipart==0.0.9
